requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
django-filter>=23.5
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
pypdf>=3.17.1
PyPDF2>=3.0.0
pdfplumber>=0.10.3
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
dj-database-url>=2.1.0
celery>=5.3.1
django-celery-results>=2.5.1
redis>=4.6.0
